Fix imports, rename packages
