--- csm_test_utils/rds_backup/fill_database/__main__.py
@@ -1,5 +1,0 @@
-#!/usr/bin/env python3
-from csm_test_utils.rds_backup.fill_database.cli import main
-
-if __name__ == '__main__':
-    main()
